Fix scraping of GitHub release page to get download URL
GitHub must have made some changes recently to release pages so that the release asset links are not part of the html from the server, but fetched using js. The js appears to request some partial html though, which we can still scrape. So now we need to make two requests instead of one to get the release asset URL.